refactor: 커스텀 인풋 / 타입별 인풋 추가 및 적용
