+++ src/refactoring/components/CartPage.tsx
@@ -1,6 +1,6 @@
 import { CartItem, Coupon, Product } from '../../types.ts';
 import { useCart } from '../hooks';
-import CustomSelect from './\bCustomSelect.tsx';
+import CustomSelect from './CustomSelect.tsx';
 import Container from './Container.tsx';
 import ContainerTitle from './ContainerTitle.tsx';
 import CustomButton from './CustomButton.tsx';
