+++ src/refactoring/components/NumberInput.tsx
@@ -1,0 +1,7 @@
+import CustomInput from './CustomInput';
+
+const NumberInput = (props: React.InputHTMLAttributes<HTMLInputElement>) => {
+  return <CustomInput {...props} type="number" />;
+};
+
+export default NumberInput;
