+++ src/refactoring/components/TextInput.tsx
@@ -1,0 +1,7 @@
+import CustomInput from './CustomInput';
+
+const TextInput = (props: React.InputHTMLAttributes<HTMLInputElement>) => {
+  return <CustomInput {...props} type="text" />;
+};
+
+export default TextInput;
